requirements.txt
#Backend
langchain>=1.1.3
langchain-openai>=1.1.1
dotenv>=0.9.9
ipykernel>=7.1.0
langchain-community>=0.4.1
mcp>=1.21.1
langchain-mcp-adapters>=0.1.13
scipy>=1.16.3
langchain-core>=1.1.3
jupyterlab>=4.5.1
pypdf>=6.2.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
requests>=2.31.0

#AI/ML
openai>=1.3.0

#Slack integration
slack-bolt==1.18.0
slackclient
slackeventsapi

#Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.32